Implement tests for the MBJSONWriter
A new interface, MBWriterTestInterface provides the methods that are different
between the JSON and XML tests. The MB{JSON,XML}WriterTestInterfaces provide
implementations of those methods.

A new abstract class AbstractMBWriterTest provides the shared implementation of
the test cases.

New classes AbstractMBWriter${entityType}Test provide an implementation of the
getDoc() method for each entity type.

Finally, MB{JSON,XML}Writer${entityType}Test classes use the common
implementation of the AbstractMBWriter${entityType} classes and the default
implementations provided by MB{JSON,XML}WriterTestInterfaces for implementing
tests with the least amount of work possible. They only need to set the corename
in their static initializer block. Inheriting this is not possible because then
each initialization of corename would only be executed once (since they're in
static initializer blocks). The corename needs to be available in a static
context because it's used in AbstractMBWriterTest.beforeClass, a static method.