Fix compat. with solr cloud
When >1 shards are used in Solr Cloud the SolrQueryResponse is retured in form of a list of SolrDocuments. This format was incompatible with the current writer causing it to error. Added the respective response parsers. This has one caveat though, the default fields to be returned now has to include _store for it to work.