SEARCH-625 (6/6):  Use MBXMLWriter own classloader
Use own plugin classloader to load JAXBContentFactory

Even after adding JAXB implementations to the library, the XML writer was unable to
load the implementations with the following error in the logs.

    Caused by: javax.xml.bind.JAXBException: Implementation of JAXB-API has not been found on module path or classpath.
     at javax.xml.bind.ContextFinder.newInstance(ContextFinder.java:278) ~[?:?]
     at javax.xml.bind.ContextFinder.find(ContextFinder.java:421) ~[?:?]
     at javax.xml.bind.JAXBContext.newInstance(JAXBContext.java:721) ~[?:?]
     at javax.xml.bind.JAXBContext.newInstance(JAXBContext.java:662) ~[?:?]
     at org.musicbrainz.search.solrwriter.MBXMLWriter.createJAXBContext(MBXMLWriter.java:280) ~[?:?]
     at org.musicbrainz.search.solrwriter.MBXMLWriter.init(MBXMLWriter.java:303) ~[?:?]
     at org.apache.solr.core.PluginBag.initInstance(PluginBag.java:104) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.PluginBag.createPlugin(PluginBag.java:142) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.PluginBag.init(PluginBag.java:277) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.PluginBag.init(PluginBag.java:266) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.SolrCore.initWriters(SolrCore.java:2730) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.SolrCore.<init>(SolrCore.java:969) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.SolrCore.<init>(SolrCore.java:874) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     at org.apache.solr.core.CoreContainer.createFromDescriptor(CoreContainer.java:1187) ~[solr-core-7.7.2.jar:7.7.2 d4c30fc2856154f2c1fefc589eb7cd070a415b94 - janhoy - 2019-05-28 23:37:48]
     ... 7 more

My first hunch was that the jar is not on the classpath. So I dug around the Solr container,
and found a directory on Solr's classpath, `/opt/solr/server/solr-webapp/webapp/WEB-INF/lib`.
Placing the jar inside this directory indeed worked. However, this is undocumented and a hack.

Looking back at the error, Solr is actually able to find `MBXMLWriter` but not the JAXB
implementation which I verified by manual inspection exists in the same jar. This points to
classloader issues, and behold ... it indeed was a classloader issue.

Reading JAXB API's code, I found it is using the current thread's context loader, printing
both in `init` method when Solr tries to load them reveals that the context loader is
different from the classloader used to load the MBXMLWriter classloader. I didn't find a
way to manually specify the classloader used by JAXB so instead setting the context
loader to plugin's classloader for the duration of initialization. And voila, that fixes
the issue.

Note by yvanzo: Solr now throws the following warning at startup:

    WARNING: An illegal reflective access operation has occurred
    WARNING: Illegal reflective access by com.sun.xml.bind.v2.runtime.reflect.opt.Injector (file:/opt/solr/lib/mb-solr-0.0.1-SNAPSHOT-jar-with-dependencies.jar) to method java.lang.ClassLoader.defineClass(java.lang.String,byte[],int,int)
    WARNING: Please consider reporting this to the maintainers of com.sun.xml.bind.v2.runtime.reflect.opt.Injector
    WARNING: Use --illegal-access=warn to enable warnings of further illegal reflective access operations
    WARNING: All illegal access operations will be denied in a future release