Correctly create and mount Solr data directories
Some directories under `/var/solr` were created in the image,
and only `/var/solr/data` was mounted to a Docker volume.

However, the directories under `/var/solr` are supposed to be created by
the script `init-var-solr`, they include logs too, and the only example
of Docker volume given in Solr documentation is bound to `/var/solr`.

This patch follows the documentation so as to simplify the Docker image,
to avoid directory ownership issues, and to persistently keep Solr logs.

Reference: https://solr.apache.org/guide/solr/9_7/deployment-guide/solr-in-docker.html#docker-compose