Copy two scripts from the Apache Solr project
These scripts are distributed within the Solr Docker image but they do
not support the SolrCloud mode. The intend of copying these is to make a
derived version to support initializing a Solr instance in SolrCloud
mode using the MusicBrainz Simple Solr Search Server Schema (mbsssss).

* start-local-solrcloud is copied from start-local-solr
* start-musicbrainz-solrcloud is copied from start-create

The license file is updated accordingly with the Apache License 2.0.