Fix permissions to fetch/remove backup archives
Previously, MusicBrainz search index dumps were fetched and removed from
the `musicbrainz` service in `musicbrainz-docker` compose project.

In order to allow the `search` service (running this image) to fetch and
remove MusicBrainz Solr backup archives, the volume has to be declared
and its mountpoint directory has to be created within the image.

The path has been intentionally changed to avoid file ownership issues
with the previous (and soon retired) volume in `musicbrainz-docker`.

References:
* https://github.com/apache/solr-docker/blob/cabde2a16c99aa974e1882df2c55ca17eab0bd17/9.7/Dockerfile#L118-L128
* https://solr.apache.org/guide/solr/9_7/deployment-guide/solr-in-docker.html#running-solr-with-host-mounted-directories