Update git submodule `mmd-schema` to tag v-2025-04-23 (#64)
Most of the changes are backwards-compatible, adding optional elements.

Only the removal of series’ ordering attribute (SEARCH-644) is a
breaking change. However, this attribute is no longer indexed since
2021, see the version 2.1.0 of SIR.

So it is safe to run with the current version 3.0.1 of SIR.

Also, the MusicBrainz Server is already using it for tests.

References:
* https://github.com/metabrainz/sir/releases/tag/v2.1.0
* https://github.com/metabrainz/musicbrainz-server/commit/7ef9e548f436e4edb37ec4262d0dd43ab4f81e3c